Remove the cmd_ prefix from command files
Now that they are in their own directory, we can remove this prefix.
This makes it easier to find a file since the prefix does not get in the
way.

Signed-off-by: Simon Glass <sjg@chromium.org>
Series-to: u-boot
Series-cc: trini
Cover-letter
Move commands into their own directory
This series makes a start at cleaning up the common directory. At present it
includes any code that is not board-specific, a library or a driver. This
means there are a lot of files, including all the commands.

Moving the command files into their own directory makes them easier to find
and is more logical than having them all in common/.
END