Makefile: Make u-boot.img the same as u-boot-dtb.img
When OF_CONTROL is enabled, u-boot.img has no purpose since it does not
include the required device tree binary. The correct image to use is
u-boot-dtb.img with OF_CONTROL but u-boot.img without OF_CONTROL.

Create u-boot.img even when OF_CONTROL is enabled, so that this file can be
used in both cases.

Signed-off-by: Simon Glass <sjg@chromium.org>