RFC: sunxi: Enable SPL FIT support
Enable SPL FIT support for the Linksprite pcDuino3 as an example of how this
feature is used.

This is only for demonstration purposes and is not to be applied.
Series-to: u-boot
Series-cc: trini, hans
Cover-letter:
spl: Support loading a FIT image containing U-Boot
We need a way to support more than one board per binary in U-Boot. Various
methods have been discussed. The one that seems to make most sense is to
adjust SPL so that it can load a FIT which contains U-Boot and several
device tree binaries. This is how things with with Linux - we load a FIT
and select the correct device tree to pass to Linux.

This series:

- Adjusts the build system to optionally build a u-boot.img in FIT format
that includes the U-Boot binary and >1 device tree files
- Adjusts SPL to support loading this
- Adds a way for SPL to determine which device tree to select (by calling a
board-specific function)
- Adjusts SPL to pass this selected device tree to U-Boot when it starts

It would be painful to require an .its file for each board just to support
this feature. In any case various people have commented that it would be
nice not to have to write this file in general. Therefore, this series
enhances mkimage to automatically generate a FIT without a .its file. So far
it understands how to add a main image and a number of device tree files. It
does not support hashing or verified boot as yet.

One problem with the FIT format as it stands is that all the data is inline.
This means that the entire file must be read in order to figure out what
device-tree files are available. It is then possible to copy the images into
place.

This is not really suitable for SPL since copying can be slow, and reading
unnecessary data would make the FIT format less efficient than the legacy
format.

Therefore this series adds a new feature to FIT which allows the images to
be stored immediately after the FIT itself ends. This makes the FIT very
small. It can be read quickly and in its entirety. Then the images can be
loaded one by one as needed. This allows SPL to support FITs containing lots
of images very efficiently.

To achieve this, mkimage is enhanced to convert between the 'normal' and
'external' version of a FIT file. The latter is only used for SPL. The main
difference is that viewing an 'external' FIT will not show the contents of
each image.

This series also includes a few other tidy-ups, such as moving mkimage's
tricky argument-processing code to use getopt().
END
Signed-off-by: Simon Glass <sjg@chromium.org>