spl: Add a way to specify a list of device trees to include
When building a FIT, more than one device tree can be included. The board
can select the one that it wants. Add a Kconfig option to allow the list of
devices trees (supported by the board) to be specified.

When using SPL_LOAD_FIT, build u-boot.img in FIT format instead of the
legacy image format. Include all the listed device tree files in this FIT.

Signed-off-by: Simon Glass <sjg@chromium.org>