usb: Correct return value in usb_stor_info()
This should return 0 on success, not 1. Fix it.

Signed-off-by: Simon Glass <sjg@chromium.org>
Series-to: u-boot
Series-cc: trini
dm: Fix various issues reported by coverity
A lot of issues have been found in the error handling with recent additions
to the FIT handling code in mkimage. In some cases buffers are not freed or
files are not closed. Fix these problems and a few others in existing code
that coverity noticed due to recent patches.
END