arm: at91: Add support for gurnard
This board is based on Snapper 9G45 which has an Atmel AT91SAM9G45 chip and
128MB of SDRAM. It includes a small LCD, 2xUSB host, SD card, Ethernet and
two UARTs.

Change-Id: I50a0de1ebd32b2e26ed7ed8d031a716a9b236229
Signed-off-by: Simon Glass <sjg@chromium.org>
Cover-letter:
at91: Convert Ethernet and LCD to driver model
This series is mainly designed to move the macb Ethernet to driver model,
along with the LCD controller. It also includes a few fixes:

- NAND with ECC doesn't build properly
- The instruction cache is off by default
- Ethernet macb doesn't send packets when the cache is enabled

A Snapper 9G45-based board is used for this work. It includes an AT91SAM9G45
CPU so is a good test of driver model on this chip.
END